✨ feat(utils/kit.ts): add utility to load @nuxt/kit with polyfill support
This new utility function `loadKit` is designed to import `@nuxt/kit` within a Nuxt project, providing a polyfill for the `writeTypes` function if the installed version of `@nuxt/schema` is below 3.7. It also includes a mechanism to resolve the correct version of Nuxt if `@nuxt/kit` is not found locally, ensuring compatibility with various Nuxt versions like `nuxt-nightly`, `nuxt3`, `nuxt`, and `nuxt-edge`. This change is necessary to maintain compatibility and ease the integration process for developers working with different versions of Nuxt in their projects.